V0.4 (#51)
* [v0.4] Edit quest screen (#44)

* make plugin dir on start and fix inserting players to db

* added access to quest editor from quests list

* added KeyHandler support when QuestBuilder made from a Quest product

* try to auto-import quests when they aren't found in the database

* SeverListener hydrating quests stored in database: Quest.fromTemplateString instead of manual.

* disabled failing on unknown fields in quest template

* use ID for quest actions instead of 'name'

* removed duplicate fields

the type field is defined with @JsonTypeInfo, so the @JsonProperty([attribute]) was not necessary

* fixed QuestRegistry.replace() not replacing database entries

* remove diary entries when quest removed (from database)

* fixed npcParticles not clearing when quests removed

* fixed unable to submit quests without a creator (universal quests)

* replaced tester quests: tina -> beans

* fixed placing Block NPCs on each other

* place NPC Block above existing instead of behind

* support for universal quests from QuestBuilder and myquests GUI

* added myquests link to unimplemented theirquest/myquest screens

* added removing all attached NPC blocks when quest removed

* added empty myquest/theirquest GUIs

* added myquest GUI title

* added theirquest GUI title

* added edit button to myquest GUI

* added back button to myquest GUI

* added back button to theirquest GUI

* added clone button to theirquest GUI

* added delete button to myquest GUI

* fixed inability to save when no NPC & action

* added confirmation for quest delete button

* fixed null action interaction errors in console

happens when interacting with an NPC with no action assigned (at that moment)

* added 'myquest' toggle button

* update version to 0.4

* data driven db migration and (unimplemented) quest toggling

* added notification for patching database

* remove and add quests on toggle

* [v0.4] Options for editing quest stage and action, including sequences/connections (#46)

* updated 'How To Get Functionality' explainers/quest action plan

* added early quest expressions/templates

* removed duplicated 'How It All Works' section

* added re-arrange buttons for actions/stages

* added ARRANGE GUI mode

currently this is unrestricted, it just lets anything be moved around anywhere in the chest inventory and in status*player inventory**git status

* disabled menu items put out of GUI (in ARRANGE mode)

* fixed taking items from GUI

fixed methods incl: throwing out of inventory, using the dragging mechanism to bypass click block (and finally) holding the block and cancelling the inventory forcing it to default to the Player inventory instead of being destroyed. These issues were fixed respectively, except the last which has a safety net sort of method, where if the GUI is closed finally, it will loop through slots and remove any items tagged with GUI="true"

* re-toggling the re-arrange button

back into the click gui mode, from arrange gui mode

* removed re-arrange buttons for stages/actions

* added 'Change Sequence' button to quest stage editor

* added a test/proof connectioneditor

* added unimplemented connectioneditor buttons

* added empty 'select connection' menu

* added 'Back' button to 'select connection' menu

* added placeholder stages for 'select connection'

* implement stages recursion in 'select connection'

* add label for 'select connection' stage buttons

* add action sub-menu outline to 'select connection'

* added back button to action sub-menu

* add actions recursion in 'select connection' sub-menu

* added setting connection data from 'connection editor'

* removed ConnectionsData debug message

* fix null getType in actiontypes/actioneditor

* added changing entry point for Quest

* fixed action in 'quest editor'

* improved 'quest stage' new action button

* fixed null diaryConnections on NPC interact

* NPC Types: Block -> (renamed to) BlockNPC

This is a breaking change sadly, but it was hell bent on not interpreting the type data under a different simple name

* added connectioneditor to actioneditor

* indicate shared/global quests in myquests

* added heading to theirquests menus

* load up questbuilder npcs with setID()

* fixed UX: some back buttons being labelled 'Exit'

* fixed unable to 'exit' in select npc block mode

* beans-tester-bonus NPCs on the same y axis

* subpar fix: changing NPC locations

* clone quest from JSON rather than former QuestBuilder class

* removed unused quest expressions

* set recovered quests to default to not shown (untoggled)

* dividers for 'queststages' to dynamically cover first two columns

* broken: added 'add stage' button

no support for multi-stage in quest diaries

* support for multi-stage in quest diaries

* add description field for GUI buttons

* Delete button on stage editor

Includes confirmation, with checks to see if possible or if stage is depended on. Shows a gray button w/ 'Cannot Delete'. Also has  a description of why, which should maybe be split on every nth whitespace.

* add a stage_0 if it doesn't already exist

* QuestBuilder.getStages() as LinkedList

* malformed quest handling improvements

still might error out to console when the error comes from parsing the JSON in Quest.fromTemplateString()

* add 'unreadable quests' indicator w/ added GUISlot.setCount

* fixed regression with cloning quests

* fixed assigning NPC block, possible race condition

usually stacks would interrupt the Listener unregistering slash cause the current re-opened GUI to dispose itself too early leading to a dead/broken GUI.

* replaced main GUI: static -> dynamic

* spring cleaning + update compiler to Java 22

* set instance on click for myquests list

* theoretical fix: GUI breaking with multiple players

* add instantiating builders map

* do not handle onClick at all for GUI if not isGUIInventory

* fix NPC interactions breaking with multiple players

multiple players interacting; or more specifically multiple diary entries with the same quest. The issue was multiple records were being obtained when trying to get the players progress in the quest from the quest diaries (this was because it was just searching by quest name, not by quest name and diary ID)

* remove static GUI (GUI template) infrastructure

removes all the JSON GUI template parsing mess

* fixed NPC 'speak' action: dialogue only showing after set

* extended NPC dialogue action, dialogue preview length

* fix empty dialogue breaking NPC speaking

* added quest action requirements + 'Speak' NPC requirement

* added NPC LocationData to validity check

* theoretical fix: add player to clearChat calls

fixes #50

* fix clearChat failing when used before function 'setUp'

* added support for more data in BlockNPC.value (half-implemented)

addresses, and basically fixes #46. But sophisticated data like custom player heads doesn't work due to a data bottleneck in NPC block selection/placement

* small fixes to quest template loading

* added catch-and-repair for missing data in myquests GUI

* fixed quest toggling removing quest from db

* added better message sending

* fixed queststage GUI behaviour when working with actions

* theoretical fix: jank when going to previous GUI from npctypes

* change some user-facing messages: ERROR -> WARN

* fixed untoggled quests adding to the QuestClient

* added PlayerQuests tutorial

* removed PLUGIN-CUTE.md template

* add the unsupported item actions to docs

* clarify 'Speak' action is not to the world

+++ src/main/java/playerquests/utility/singleton/QuestRegistry.java
@@ -1,25 +1,29 @@
 package playerquests.utility.singleton;
 
+import java.io.IOException;
 import java.util.HashMap; // hash table map
+import java.util.List;
 import java.util.Map; // generic map type
+import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean; // modify boolean state in a stream operation
+import java.util.stream.Collectors;
 
 import org.bukkit.Bukkit; // accessing Bukkit API
 import org.bukkit.entity.HumanEntity; // representing players and other humanoid entities
 import org.bukkit.entity.Player; // representing players
 
+import playerquests.builder.quest.data.LocationData;
 import playerquests.builder.quest.npc.QuestNPC; // describes quest NPCs
 import playerquests.client.quest.QuestClient; // player quest state
 import playerquests.product.Quest; // describes quests
 import playerquests.utility.ChatUtils; // utility methods related to chat
+import playerquests.utility.FileUtils;
+import playerquests.utility.ChatUtils.MessageType;
 
 
 /**
  * Singleton for putting and accessing quest products from anywhere.
  */
-// TODO: event listeners?
-// TODO: (on remove) warn if someone is currently playing the quest
-// TODO: (on remove) if called twice, forces removal even if someone is playing the quest (deletion waiting list)
-// TODO: (on quest completion) tell quest creators in deletion waiting list the quest is safe to remove
 public class QuestRegistry {
     
     /**
@@ -36,6 +40,11 @@ public class QuestRegistry {
      * The map holding the quest clients (questers).
      */
     private final Map<Player, QuestClient> questers = new HashMap<Player, QuestClient>();
+
+    /**
+     * The resource folder quests are in
+     */
+    private final String questPath = "quest/templates";
 
     /**
      * Private constructor to prevent instantiation.
@@ -58,6 +67,9 @@ public class QuestRegistry {
         // add to the registry map
         registry.put(quest.getID(), quest);
 
+        // store ref to database
+        Database.addQuest(quest.getID());
+
         // place the NPCs in the world
         quest.getNPCs().entrySet().stream()
             .forEach(entry -> {
@@ -76,8 +88,15 @@ public class QuestRegistry {
      */
     public void submit(Quest quest) {
         String questID = quest.getID();
-        HumanEntity creator = Bukkit.getPlayer(quest.getCreator());
+        HumanEntity creator = null;
+        AtomicBoolean isQuestValid = new AtomicBoolean(true); // start with assumption the quest is valid
+        
+        // get the creator if it's not a universal quest (missing or null creator)
+        if (quest.getCreator() != null) {
+            creator = Bukkit.getPlayer(quest.getCreator());
+        }
 
+        // replace if already found in registry
         if (registry.get(questID) != null) {
             if (creator != null) {
                 creator.sendMessage("[Updating the quest]");
@@ -86,10 +105,42 @@ public class QuestRegistry {
             return;
         }
 
-        // store ref to database
-        Database.addQuest(questID);
+        // load up a list of NPC locations to help figure if an npc already exists at the npc.location
+        List<LocationData> registryNPCLocations = this.registry.values().stream().flatMap(
+            registryQuest -> registryQuest.getNPCs().values().stream().map(QuestNPC::getLocation)
+        ).collect(Collectors.toList());
 
-        // store ref to registry
+        // check each registry NPC location against current NPC in submitted quest
+        Optional<QuestNPC> collidingNPC = quest.getNPCs().values().stream()
+            .filter(questNPC -> registryNPCLocations.stream()
+                .anyMatch(existingLocation -> {
+                    LocationData submittedLocation = questNPC.getLocation();
+
+                    if (!questNPC.isValid()) {
+                        isQuestValid.set(false);
+                        return false; // invalid, so no use displacing
+                    }
+
+                    return existingLocation.collidesWith(submittedLocation);
+                })
+            ).findFirst();
+
+        // do not continue if npc has no LocationData
+        if (!isQuestValid.get()) {
+            return;
+        }
+
+        // if submitted quest has an NPC found colliding with existing,
+        if (collidingNPC.isPresent()) {
+            LocationData questNPCLocation = collidingNPC.get().getLocation();
+            questNPCLocation.setY(questNPCLocation.getY() + 1); // put the NPC above the existing one
+            this.submit(quest); // resubmit the quest to test if new location collides
+
+            // do not continue if an npc already exists at the crucial npc.location
+            return;
+        }
+
+        // store ref
         this.add(quest);
 
         if (creator != null) {
@@ -102,14 +153,54 @@ public class QuestRegistry {
      * @param quest the quest to remove.
      */
     public void remove(Quest quest) {
-        // remove ref from database
-        Database.removeQuest(quest.getID());
+        this.remove(quest, false);
+    }
 
+    /**
+     * Removes a quest from the registry.
+     * @param quest the quest to remove
+     * @param preserveInDatabase whether to keep untoggled in db; just remove from world
+     */
+    public void remove(Quest quest, Boolean preserveInDatabase) {
+        // remove ref from database
+        if (!preserveInDatabase) {
+            Database.removeQuest(quest.getID());
+        }
+
+        // remove ref from registry
         registry.remove(quest.getID());
 
+        // remove ref from questers
         questers.values().stream().forEach(quester -> {
             quester.removeQuest(quest);
         });
+
+        // remove traces from world
+        PlayerQuests.getInstance().remove(quest);
+    }
+
+    /**
+     * Deletes a quest from the plugin.
+     * @param quest the quest to delete
+     * @return whether the operation was successful or not
+     */
+    public Boolean delete(Quest quest) {
+        // try to remove from files
+        try {
+            FileUtils.delete(this.questPath + "/" + quest.getID() + ".json");
+            
+            // remove from registry
+            this.remove(quest);
+
+        } catch (IOException e) {
+            ChatUtils.message("Could not delete the " + quest.getTitle() + " quest")
+                .player(Bukkit.getPlayer(quest.getCreator()))
+                .type(MessageType.ERROR)
+                .send();
+            return false;
+        }
+
+        return true;
     }
 
     /**
@@ -151,5 +242,38 @@ public class QuestRegistry {
     public void clear() {
         this.registry.clear();
         this.questers.clear();
+    }
+
+    /**
+     * Get a quest from the quest registry.
+     * If fails from quest registry, it will
+     * search the questPath (resources folder).
+     * @param questID the quest ID
+     * @return the quest object
+     */
+    public Quest getQuest(String questID) {
+        Quest result = this.getAllQuests().get(questID);
+
+        // search in filesystem
+        if (result == null) {
+            System.err.println("Quest registry could not find quest: " + questID + ". It'll now search for it in the resources quest template files.");
+
+            // attempt finding it in the files and uploading to database
+            try {
+                result = Quest.fromTemplateString(FileUtils.get(this.questPath + "/" + questID + ".json"));
+
+                // if everything is okay, submit the quest to the database 
+                // to avoid having to search for it again like this.
+                if (result != null) {
+                    this.submit(result);
+                    result.toggle(false); // toggle off when found (opposite of default)
+                }
+
+            } catch (IOException e) {
+                throw new RuntimeException("Could not find the quest with ID: " + questID, e);
+            }
+        }
+
+        return result;
     }
 }
